+++ DESIRE-Engine/src/Core/math/vectormath/Vector4.h
@@ -153,7 +153,7 @@ public:
 	// Compute the length of a 4-D vector
 	DESIRE_FORCE_INLINE float Length() const
 	{
-		return std::sqrtf(LengthSqr());
+		return std::sqrt(LengthSqr());
 	}
 
 	// Normalize a 4-D vector
