+++ UnitTest/src/Core/math/test_Vector4.cpp
@@ -127,7 +127,7 @@ TEST_CASE("Vector4", "[Core][math]")
 
 	SECTION("Length()")
 	{
-		CHECK(vec0.Length() == Approx(std::sqrtf(vec0.LengthSqr())));
+		CHECK(vec0.Length() == Approx(std::sqrt(vec0.LengthSqr())));
 	}
 
 	SECTION("Normalize() | Normalized()")
